Add Global Top 10 Leaderboard (#111)
* Refactor leaderboard owner to have normal membership

* Fix deprecation warning

* Remove inspect

* Add local IP addresses to CORS for dev

* Add results to `leaderboards#index`

* Setting policy to true since owner will now have a membership

* Fix leaderboards association

* Fix leaderboard policy

* Create scenic views

* Make views materialized

* Fix queries in views

* Update views and associations

* Fix match results view

* Refactor matches controller

* Add missing lines at EOF

* Fix leaderboard rankings and remove unnecessary associations

* Clean up unused code

* Refactor views into partials

* Add accuracy to users#show response

* Add missing lines at EOF

* Add refresh views callback to most models

* Add migration and rake task to seed

* Add global leaderboards

* Add leaderboard description and change name

* Create service to deactivate callback on bulk update operations

* Move auto join logic to User after_create callback

+++ app/controllers/v1/leaderboards_controller.rb
@@ -22,6 +22,7 @@ class V1::LeaderboardsController < ApplicationController
     @leaderboard = Leaderboard.find(params[:id])
     authorize @leaderboard
     membership = @leaderboard.memberships.find_by!(user: current_user)
+    authorize membership
     membership.destroy
     head :no_content
   end
